+++ packages/codeceptjs-configure/index.js
@@ -4,13 +4,14 @@ const configure = require('./lib/config/configure');
 const host = require('./lib/host/host');
 const logger = require('./lib/logger/logger');
 const reportCollector = require('./lib/reports/report.collector').collect;
-
+const allureReportLauncher = require('./lib/reports/allure-static-launcher/allure-launcher').allureReportLauncher;
 const Driver_commands = require.resolve('./lib/helpers/driver-commands.helper.js');
 
 module.exports = {
     config: { master },
     helpers: { Driver_commands },
     cleanReports,
+    allureReportLauncher,
     reportCollector,
     configure,
     host,